merge-cls

+++ src/main/java/com/spring3/oauth/jwt/models/request/UpsertNovelRequest.java
@@ -1,22 +1,13 @@
 package com.spring3.oauth.jwt.models.request;
 
-import com.spring3.oauth.jwt.entity.Author;
-import com.spring3.oauth.jwt.entity.Genre;
 import com.spring3.oauth.jwt.entity.enums.NovelStatusEnum;
-import jakarta.persistence.Column;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToMany;
-import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.List;
-
-import static jakarta.persistence.FetchType.LAZY;
 
 @Data
 @AllArgsConstructor
@@ -33,6 +24,6 @@ public class UpsertNovelRequest {
     private int totalChapters;
     private BigDecimal averageRatings;
     private int likeCounts;
-    private Integer authorId;
+    private long authorId;
     private List<Integer> genreIds;
 }
